airflow/requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
exceptiongroup==1.0.0rc9
fake-headers==1.0.2
h11==0.14.0
html5lib==1.1
idna==3.4
numpy==1.21.6
outcome==1.2.0
pandas
psycopg2-binary==2.9.4
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
webencodings==0.5.1
wsproto==1.2.0
simplejson==3.17.6
dateparser